Kernel/Storage: Move ATA device signature definitions to a general file
