+++ Kernel/Storage/ATA/AHCI/Port.h
@@ -23,6 +23,7 @@
 #include <Kernel/Storage/ATA/AHCI/Definitions.h>
 #include <Kernel/Storage/ATA/AHCI/InterruptHandler.h>
 #include <Kernel/Storage/ATA/ATADevice.h>
+#include <Kernel/Storage/ATA/Definitions.h>
 #include <Kernel/WaitQueue.h>
 
 namespace Kernel {
@@ -41,7 +42,7 @@ public:
     u32 port_index() const { return m_port_index; }
     u32 representative_port_index() const { return port_index() + 1; }
     bool is_operable() const;
-    bool is_atapi_attached() const { return m_port_registers.sig == (u32)AHCI::DeviceSignature::ATAPI; };
+    bool is_atapi_attached() const { return m_port_registers.sig == (u32)ATA::DeviceSignature::ATAPI; };
 
     RefPtr<StorageDevice> connected_device() const { return m_connected_device; }
